Task: add Zscore , Population Standard Deviation, Variance of population proportion

+++ Statistics/Median.py
@@ -2,6 +2,6 @@ def median(pool):
     copy = sorted(pool)
     size = len(copy)
     if size % 2 == 1:
-        return copy[(size - 1) / 2]
+        return int(copy[int((size - 1) / 2)])
     else:
-        return (copy[size / 2 - 1] + copy[size / 2]) / 2
+        return int((copy[int(size / 2 - 1)] + copy[int(size / 2)]) / 2)
